Reorganize Viewer implementation folders
Copilot-Session: b3644529-ee71-4272-b6fc-cf5d0862b979
