+++ packages/public/@babylonjs/viewer/package.json
@@ -2,14 +2,14 @@
     "name": "@babylonjs/viewer",
     "version": "9.16.2",
     "type": "module",
-    "main": "lib/index.js",
-    "module": "lib/index.js",
-    "esnext": "lib/index.js",
-    "types": "lib/index.d.ts",
+    "main": "lib/full/index.js",
+    "module": "lib/full/index.js",
+    "esnext": "lib/full/index.js",
+    "types": "lib/full/index.d.ts",
     "exports": {
         ".": {
-            "types": "./lib/index.d.ts",
-            "default": "./lib/index.js"
+            "types": "./lib/full/index.d.ts",
+            "default": "./lib/full/index.js"
         },
         "./lite": {
             "types": "./lib/lite/index.d.ts",
