+++ packages/tools/viewer-configurator/src/components/babylonViewer/viewer.tsx
@@ -1,9 +1,9 @@
-import { type ViewerElement, type ViewerOptions } from "viewer/index";
+import { type ViewerElement, type ViewerOptions } from "viewer/full/index";
 
 import { makeStyles } from "@fluentui/react-components";
 import { Logger } from "core/Misc/logger";
 import { useEffect, type FunctionComponent } from "react";
-import { ConfigureCustomViewerElement } from "viewer/viewerElement";
+import { ConfigureCustomViewerElement } from "viewer/full/viewerElement";
 import "viewer";
 
 const useStyles = makeStyles({
