+++ packages/tools/viewer-configurator/src/modelLoader.ts
@@ -1,5 +1,5 @@
 import { type Nullable } from "core/index";
-import { type LoadModelOptions, type ViewerElement } from "viewer/index";
+import { type LoadModelOptions, type ViewerElement } from "viewer/full/index";
 
 import { Deferred } from "core/Misc/deferred";
 import { GetRegisteredSceneLoaderPluginMetadata } from "core/Loading/sceneLoader";
