+++ packages/tools/viewer/src/full/Shaders/envShadowGround.fragment.fx
