+++ packages/tools/viewer/src/full/Shaders/envShadowGround.vertex.fx
