+++ packages/tools/viewer/src/full/ShadersWGSL/envShadowGround.fragment.fx
