+++ packages/tools/viewer/src/full/ShadersWGSL/envShadowGround.vertex.fx
