+++ packages/tools/viewer/src/full/index.ts
@@ -6,13 +6,13 @@
 // re-exported viewer modules pull in any decorated core class.
 import "core/Misc/symbolMetadataPolyfill";
 
-export type { CameraAutoOrbit, EnvironmentOptions, HotSpot, PostProcessing, ShadowQuality, ToneMapping, ViewerHotSpotQuery } from "./viewerBase";
+export type { CameraAutoOrbit, EnvironmentOptions, HotSpot, PostProcessing, ShadowQuality, ToneMapping, ViewerHotSpotQuery } from "../viewerBase";
 export type { LoadModelOptions, Model, ViewerDetails, ViewerOptions } from "./viewer";
 export type { CanvasViewerOptions } from "./viewerFactory";
-export type { ViewerElementEventMap } from "./viewerElementBase";
+export type { ViewerElementEventMap } from "../viewerElementBase";
 
-export { ViewerHotSpotResult } from "./viewerBase";
+export { ViewerHotSpotResult } from "../viewerBase";
 export { CreateHotSpotFromCamera, DefaultViewerOptions, Viewer } from "./viewer";
 export { ConfigureCustomViewerElement, HTML3DElement, ViewerElement } from "./viewerElement";
 export { CreateViewerForCanvas } from "./viewerFactory";
-export { HTML3DAnnotationElement } from "./viewerAnnotationElement";
+export { HTML3DAnnotationElement } from "../viewerAnnotationElement";
