+++ packages/tools/viewer/src/lite/index.ts
@@ -1,8 +1,8 @@
 export type { CameraAutoOrbit, EnvironmentOptions, HotSpot, PostProcessing, ShadowQuality, ToneMapping, ViewerHotSpotQuery } from "../viewerBase";
-export type { CanvasViewerOptions, LoadModelOptions, ViewerOptions } from "../viewerLite";
+export type { CanvasViewerOptions, LoadModelOptions, ViewerOptions } from "./viewer";
 export type { ViewerElementEventMap } from "../viewerElementBase";
 
 export { ViewerHotSpotResult } from "../viewerBase";
-export { ViewerElement, HTML3DElement, ConfigureCustomViewerElement } from "../viewerElementLite";
-export { Viewer, CreateViewerForCanvas, DefaultViewerOptions } from "../viewerLite";
+export { ViewerElement, HTML3DElement, ConfigureCustomViewerElement } from "./viewerElement";
+export { Viewer, CreateViewerForCanvas, DefaultViewerOptions } from "./viewer";
 export { HTML3DAnnotationElement } from "../viewerAnnotationElement";
