+++ packages/tools/viewer/test/apps/web/analyze.html
@@ -23,6 +23,6 @@
             environment="../../../../../public/@babylonjs/viewer/assets/photoStudio.env"
         >
         </babylon-viewer>
-        <script type="module" src="/packages/tools/viewer/dist/analyze/tools/viewer/src/index.js"></script>
+        <script type="module" src="/packages/tools/viewer/dist/analyze/tools/viewer/src/full/index.js"></script>
     </body>
 </html>
