+++ packages/tools/viewer/test/apps/web/coverage.html
@@ -24,7 +24,7 @@
         >
         </babylon-viewer>
         <button id="saveCoverageButton" style="position: absolute; top: 10px; left: 10px" onclick="saveCoverage()">Collect Coverage</button>
-        <script type="module" src="/packages/tools/viewer/dist/coverage/instrumented/tools/viewer/src/index.js"></script>
+        <script type="module" src="/packages/tools/viewer/dist/coverage/instrumented/tools/viewer/src/full/index.js"></script>
         <script>
             async function saveCoverage() {
                 const coverageButton = document.getElementById("saveCoverageButton");
