+++ packages/tools/viewer/test/apps/web/test.html
@@ -20,6 +20,6 @@
             }
         </style>
 
-        <script type="module" src="/packages/tools/viewer/src/index.ts"></script>
+        <script type="module" src="/packages/tools/viewer/src/full/index.ts"></script>
     </head>
 </html>
